Bug 13293 - Regression: Override renewal limit option broken with AJAX circ
The "Override renewal limit" checkbox no longer works to un-hide the
renewal checkboxes in the table of checkouts. I assume this is because
the script works in such a way that it is looking for the checkboxes
before they are part of the document.

Test Plan:
1) Apply this patch
2) Check the "Override renewal limit" checkbox
3) Note the checkbox now appears

Note, this change allows the "X of Y renewals remaining" to be displayed
for non-renewable items. I left this as it is for two reasons: 1) it
gives the librarian more information that may be useful, and 2) it looks
nicer and more uniform.

Signed-off-by: Frederic Demians <f.demians@tamil.fr>

I confirm the bug and the solution.

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Passes tests and QA script, works as described.
Exception: on-site checkouts can't be renewed.

Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>

Conflicts:
	koha-tmpl/intranet-tmpl/prog/en/js/checkouts.js